move ascwds worker
