+++ projects/_01_ingest/unittest_data/ingest_test_file_schemas.py
@@ -180,6 +180,60 @@ class ASCWDSWorkplaceSchemas:
 
 
 @dataclass
+class ASCWDSWorkerSchemas:
+    worker_schema = StructType(
+        [
+            StructField(AWK.location_id, StringType(), True),
+            StructField(AWK.establishment_id, StringType(), True),
+            StructField(AWK.worker_id, StringType(), True),
+            StructField(AWK.main_job_role_id, StringType(), True),
+            StructField(AWK.import_date, StringType(), True),
+            StructField(AWK.year, StringType(), True),
+            StructField(AWK.month, StringType(), True),
+            StructField(AWK.day, StringType(), True),
+        ]
+    )
+
+    create_clean_main_job_role_column_schema = StructType(
+        [
+            StructField(AWKClean.worker_id, StringType(), True),
+            StructField(AWKClean.ascwds_worker_import_date, DateType(), True),
+            StructField(AWKClean.main_job_role_id, StringType(), True),
+        ]
+    )
+    expected_create_clean_main_job_role_column_schema = StructType(
+        [
+            *create_clean_main_job_role_column_schema,
+            StructField(AWKClean.main_job_role_clean, StringType(), True),
+            StructField(AWKClean.main_job_role_clean_labelled, StringType(), True),
+        ]
+    )
+
+    replace_care_navigator_with_care_coordinator_schema = StructType(
+        [
+            StructField(AWKClean.worker_id, StringType(), True),
+            StructField(AWKClean.main_job_role_clean, StringType(), True),
+        ]
+    )
+
+    impute_not_known_job_roles_schema = StructType(
+        [
+            StructField(AWKClean.worker_id, StringType(), True),
+            StructField(AWKClean.ascwds_worker_import_date, DateType(), True),
+            StructField(AWKClean.main_job_role_clean, StringType(), True),
+        ]
+    )
+
+    remove_workers_with_not_known_job_role_schema = StructType(
+        [
+            StructField(AWKClean.worker_id, StringType(), True),
+            StructField(AWKClean.ascwds_worker_import_date, DateType(), True),
+            StructField(AWKClean.main_job_role_clean, StringType(), True),
+        ]
+    )
+
+
+@dataclass
 class CapacityTrackerCareHomeSchema:
     sample_schema = StructType(
         [
