+++ terraform/pipeline/glue.tf
@@ -97,7 +97,7 @@ module "ingest_ascwds_dataset_job" {
 
 module "clean_ascwds_worker_job" {
   source            = "../modules/glue-job"
-  script_dir        = "jobs"
+  script_dir        = "projects/_01_ingest/ascwds/jobs"
   script_name       = "clean_ascwds_worker_data.py"
   glue_role         = aws_iam_role.sfc_glue_service_iam_role
   worker_type       = "G.1X"
